Rename some nihms-submission modules and classes (#38)
* Rename nihms-model to deposit-model

With the decision to use a single model for all deposit service targets,
it is appropriate to rename the existing model.  The model had been
created when initial development only targeted a NIHMS server, and so
that name had been used.  Soon, the model will also be used when
targeting a DSpace server, and so a more generic name is now preferable.

This commit renames the model project and its classes and updates
dependencies and references throughout the nihms-submission module.
Later renaming efforts will address other projects and ultimately change
the package groupId.

A few additional edits are included, and several references to NIHMS
were let in place until their appropriateness can be further evaluated.

* Builder modules renamed to be target-agnostic

Since the deposit service data model will be used by all targets, the
modules that define how the models are built should have target-agnostic
names.  Here, nihms-builder is renamed to builder-api (to match the
naming for the assembler and transport api modules) and
nihms-filesystem-builder is renamed to filesystem-builder.  Shortly, a
new fcrepo-builder module will be added to build models from data stored
in a Fedora repository.

* Commit ommitted file

This file should have been included in the previous commit.

* Add one more POM change related to module renaming.

* Alter a previous renaming of what was nihms-filesystem-builder

Originally, nihms-filesystem-builder was renamed to filesystem-builder,
but as I was working to reconcile that with the need for very similar
functionality to read from a Fedora repository, it became clear that
instead of having three projects (one to copy data and two derived from
it for filesystem and Fedora data), that this should all be done in one
module with three classes for the separate functionality.  Then, other
modules derived from builder-api would be for data that came from places
other than Fedora.
