+++ builder-api/pom.xml
@@ -25,20 +25,20 @@
         <version>1.0.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>nihms-transport-api</artifactId>
-    <name>Transport API</name>
+    <artifactId>builder-api</artifactId>
+    <name>Submission Model Builder</name>
 
     <dependencies>
 
         <dependency>
-            <groupId>org.dataconservancy.nihms</groupId>
-            <artifactId>nihms-assembler-api</artifactId>
-            <version>${project.parent.version}</version>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
         </dependency>
 
         <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
+            <groupId>org.dataconservancy.nihms</groupId>
+            <artifactId>deposit-model</artifactId>
+            <version>${project.parent.version}</version>
         </dependency>
 
     </dependencies>
