+++ deposit-model/pom.xml
@@ -25,22 +25,7 @@
         <version>1.0.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>nihms-transport-api</artifactId>
-    <name>Transport API</name>
-
-    <dependencies>
-
-        <dependency>
-            <groupId>org.dataconservancy.nihms</groupId>
-            <artifactId>nihms-assembler-api</artifactId>
-            <version>${project.parent.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-        </dependency>
-
-    </dependencies>
+    <artifactId>deposit-model</artifactId>
+    <name>Submission Model</name>
 
 </project>
