+++ deposit-model/src/main/java/org/dataconservancy/nihms/model/package-info.java
