+++ deposit-model/src/main/resources/nihms-bulksub-specs.dtd
