+++ fedora-builder/src/main/java/org/dataconservancy/nihms/builder/fs/NihmsBuilderPropertyNames.java
