+++ nihms-assembler-api/src/main/java/org/dataconservancy/nihms/assembler/PackageStream.java
@@ -15,12 +15,11 @@
  */
 package org.dataconservancy.nihms.assembler;
 
-import org.dataconservancy.nihms.model.NihmsSubmission;
+import org.dataconservancy.nihms.model.DepositSubmission;
 
 import java.io.InputStream;
 import java.util.Collection;
 import java.util.Iterator;
-import java.util.Map;
 
 /**
  * A streamable serialized form of a submission package.
@@ -81,7 +80,7 @@ public interface PackageStream {
 
         /**
          * A suggested name for this package.  The {@link #spec() specification} used for
-         * {@link Assembler#assemble(NihmsSubmission) assembling} a package may place requirements on the name of the
+         * {@link Assembler#assemble(DepositSubmission) assembling} a package may place requirements on the name of the
          * package file in the target system.  For example, BagIt recommends that the name of the package file be based
          * on the name of the base directory of the bag.  Submission components responsible for streaming {@link
          * PackageStream this package} to target systems can use the name returned by this method as the name of the
