+++ nihms-cli/pom.xml
@@ -68,7 +68,7 @@
 
         <dependency>
             <groupId>org.dataconservancy.nihms</groupId>
-            <artifactId>nihms-filesystem-builder</artifactId>
+            <artifactId>fedora-builder</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
