+++ nihms-submission-engine/pom.xml
@@ -37,7 +37,7 @@
 
         <dependency>
             <groupId>org.dataconservancy.nihms</groupId>
-            <artifactId>nihms-model</artifactId>
+            <artifactId>deposit-model</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
 
@@ -49,7 +49,7 @@
 
         <dependency>
             <groupId>org.dataconservancy.nihms</groupId>
-            <artifactId>nihms-builder</artifactId>
+            <artifactId>fedora-builder</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
