+++ pom.xml
@@ -25,9 +25,9 @@
     <name>NIH Package and Submission Parent Project</name>
 
     <modules>
-        <module>nihms-model</module>
-        <module>nihms-builder</module>
-        <module>nihms-filesystem-builder</module>
+        <module>deposit-model</module>
+        <module>builder-api</module>
+        <module>fedora-builder</module>
         <module>nihms-assembler-api</module>
         <module>nihms-native-assembler</module>
         <!-- nihms-bagit-assembler -->
